chore(suite-types): move guide types from suite-data to suite-types
