+++ packages/suite-desktop/package.json
@@ -185,6 +185,7 @@
     },
     "dependencies": {
         "@sentry/electron": "3.0.0",
+        "@suite-common/suite-types": "*",
         "@suite-common/suite-utils": "*",
         "@trezor/connect": "9.0.2",
         "@trezor/connect-common": "*",
