+++ suite-common/suite-types/package.json
@@ -13,16 +13,10 @@
         "@reduxjs/toolkit": "^1.8.3",
         "@suite-common/metadata-types": "*",
         "@suite-common/wallet-config": "*",
-        "@trezor/connect": "9.0.2",
-        "@trezor/suite-data": "*"
+        "@trezor/connect": "9.0.2"
     },
     "devDependencies": {
         "jest": "^26.6.3",
         "typescript": "4.7.4"
-    },
-    "nx": {
-        "implicitDependencies": [
-            "@trezor/suite-data"
-        ]
     }
 }
